Split server methods out of Project module
