+++ packages/xdl/src/Project.ts
@@ -10,43 +10,31 @@ import {
   PackageJSONConfig,
   Platform,
   ProjectTarget,
-  readExpRcAsync,
   resolveModule,
 } from '@expo/config';
-import { getBareExtensions, getManagedExtensions } from '@expo/config/paths';
-import { bundleAsync, MetroDevServerOptions, runMetroDevServerAsync } from '@expo/dev-server';
+import { bundleAsync } from '@expo/dev-server';
 import JsonFile from '@expo/json-file';
 import joi from '@hapi/joi';
 import assert from 'assert';
 import axios from 'axios';
 import chalk from 'chalk';
-import child_process from 'child_process';
 import crypto from 'crypto';
 import decache from 'decache';
-import delayAsync from 'delay-async';
-import express from 'express';
 import FormData from 'form-data';
-import freeportAsync from 'freeport-async';
 import fs from 'fs-extra';
 import getenv from 'getenv';
 import HashIds from 'hashids';
-import escapeRegExp from 'lodash/escapeRegExp';
-import { AddressInfo } from 'net';
 import path from 'path';
 import readLastLines from 'read-last-lines';
 import semver from 'semver';
 import slug from 'slugify';
-import split from 'split';
-import treekill from 'tree-kill';
 import urljoin from 'url-join';
-import { promisify } from 'util';
 import uuid from 'uuid';
 
 import Analytics from './Analytics';
 import * as Android from './Android';
 import ApiV2 from './ApiV2';
 import Config from './Config';
-import * as ConnectionStatus from './ConnectionStatus';
 import * as DevSession from './DevSession';
 import * as EmbeddedAssets from './EmbeddedAssets';
 import { maySkipManifestValidation } from './Env';
@@ -59,22 +47,25 @@ import * as ThirdParty from './ThirdParty';
 import * as UrlUtils from './UrlUtils';
 import UserManager, { ANONYMOUS_USERNAME, User } from './User';
 import * as Versions from './Versions';
-import * as Watchman from './Watchman';
 import * as Webpack from './Webpack';
 import XDLError from './XDLError';
 import * as ExponentTools from './detach/ExponentTools';
 import * as TableText from './logs/TableText';
 import { learnMore } from './logs/TerminalLink';
 import * as Doctor from './project/Doctor';
-import { getManifestHandler } from './project/ManifestHandler';
 import * as ProjectUtils from './project/ProjectUtils';
 import { assertValidProjectRoot } from './project/errors';
 import { startTunnelsAsync, stopTunnelsAsync } from './project/ngrok';
+import { startDevServerAsync, StartOptions } from './project/startDevServerAsync';
+import { startExpoServerAsync, stopExpoServerAsync } from './project/startLegacyExpoServerAsync';
+import {
+  startReactNativeServerAsync,
+  stopReactNativeServerAsync,
+} from './project/startLegacyReactNativeServerAsync';
 import { writeArtifactSafelyAsync } from './tools/ArtifactUtils';
 import { resolveEntryPoint } from './tools/resolveEntryPoint';
-const MINIMUM_BUNDLE_SIZE = 500;
 
-const treekillAsync = promisify<number, string>(treekill);
+const MINIMUM_BUNDLE_SIZE = 500;
 
 type SelfHostedIndex = ExpoAppManifest & {
   dependencies: string[];
@@ -82,16 +73,6 @@ type SelfHostedIndex = ExpoAppManifest & {
 
 type LoadedHook = Hook & {
   _fn: (input: HookArguments) => any;
-};
-
-export type StartOptions = {
-  devClient?: boolean;
-  reset?: boolean;
-  nonInteractive?: boolean;
-  nonPersistent?: boolean;
-  maxWorkers?: number;
-  webOnly?: boolean;
-  target?: ProjectTarget;
 };
 
 type PublishOptions = {
@@ -145,15 +126,6 @@ export async function currentStatus(projectDir: string): Promise<ProjectStatus> 
   } else {
     return 'exited';
   }
-}
-
-async function _getFreePortAsync(rangeStart: number) {
-  const port = await freeportAsync(rangeStart, { hostnames: [null, 'localhost'] });
-  if (!port) {
-    throw new XDLError('NO_PORT_FOUND', 'No available port found');
-  }
-
-  return port;
 }
 
 function _requireFromProject(modulePath: string, projectRoot: string, exp: ExpoConfig) {
@@ -1301,501 +1273,6 @@ export async function startBuildAsync(
   return await api.putAsync('build/start', { manifest: exp, options });
 }
 
-async function _waitForRunningAsync(
-  projectRoot: string,
-  url: string,
-  retries: number = 300
-): Promise<true> {
-  try {
-    const response = await axios.request({
-      url,
-      responseType: 'text',
-      proxy: false,
-    });
-    if (/packager-status:running/.test(response.data)) {
-      return true;
-    } else if (retries === 0) {
-      ProjectUtils.logError(
-        projectRoot,
-        'expo',
-        `Could not get status from Metro bundler. Server response: ${response.data}`
-      );
-    }
-  } catch (e) {
-    if (retries === 0) {
-      ProjectUtils.logError(
-        projectRoot,
-        'expo',
-        `Could not get status from Metro bundler. ${e.message}`
-      );
-    }
-  }
-
-  if (retries <= 0) {
-    throw new Error('Connecting to Metro bundler failed.');
-  } else {
-    await delayAsync(100);
-    return _waitForRunningAsync(projectRoot, url, retries - 1);
-  }
-}
-
-// The --verbose flag is intended for react-native-cli/metro, not expo-cli
-const METRO_VERBOSE_WARNING = 'Run CLI with --verbose flag for more details.';
-
-// Remove these constants and related code when SDK35 isn't supported anymore
-// Context: https://github.com/expo/expo-cli/issues/1074
-const NODE_12_WINDOWS_METRO_ERROR = `Invalid regular expression: /(.*\\__fixtures__\\.*|node_modules[\\]react[\\]dist[\\].*|website\\node_modules\\.*|heapCapture\\bundle.js|.*\\__tests__\\.*)$/: Unterminated character class`;
-const NODE_12_WINDOWS_METRO_SUGGESTION = `\nUnable to start the project due to a documented incompatibility between Node 12 LTS and Expo SDK 35 on Windows.
-Please refer to this GitHub comment for a solution:
-https://github.com/expo/expo-cli/issues/1074#issuecomment-559220752\n`;
-
-function _logPackagerOutput(projectRoot: string, level: string, data: object) {
-  let output = data.toString();
-  if (!output) {
-    return;
-  }
-  // Temporarily hide warnings about duplicate providesModule declarations
-  // under react-native
-  if (_isIgnorableDuplicateModuleWarning(projectRoot, level, output)) {
-    ProjectUtils.logDebug(
-      projectRoot,
-      'expo',
-      `Suppressing @providesModule warning: ${output}`,
-      'project-suppress-providesmodule-warning'
-    );
-    return;
-  }
-  if (_isIgnorableMetroConsoleOutput(output) || _isIgnorableRnpmWarning(output)) {
-    ProjectUtils.logDebug(projectRoot, 'expo', output);
-    return;
-  }
-
-  if (output.includes(NODE_12_WINDOWS_METRO_ERROR)) {
-    ProjectUtils.logError(projectRoot, 'expo', NODE_12_WINDOWS_METRO_SUGGESTION);
-    return;
-  }
-
-  if (output.includes(METRO_VERBOSE_WARNING)) {
-    output = output.replace(METRO_VERBOSE_WARNING, '');
-  }
-
-  if (/^Scanning folders for symlinks in /.test(output)) {
-    ProjectUtils.logDebug(projectRoot, 'metro', output);
-    return;
-  }
-  if (level === 'info') {
-    ProjectUtils.logInfo(projectRoot, 'metro', output);
-  } else {
-    ProjectUtils.logError(projectRoot, 'metro', output);
-  }
-}
-
-function _isIgnorableMetroConsoleOutput(output: string) {
-  // As of react-native 0.61.x, Metro prints console logs from the device to console, without
-  // passing them through the custom log reporter.
-  //
-  // Managed apps have a separate remote logging implementation included in the Expo SDK,
-  // (see: _handleDeviceLogs), so we can just ignore these device logs from Metro.
-  // if (/^ () /)
-  //
-  // These logs originate from:
-  // https://github.com/facebook/metro/blob/e8181fb9db7db31adf7d1ed9ab840f54449ef238/packages/metro/src/lib/logToConsole.js#L50
-  return /^\s+(INFO|WARN|LOG|GROUP|DEBUG) /.test(output);
-}
-
-function _isIgnorableRnpmWarning(output: string) {
-  return output.startsWith(
-    'warn The following packages use deprecated "rnpm" config that will stop working from next release'
-  );
-}
-
-function _isIgnorableDuplicateModuleWarning(
-  projectRoot: string,
-  level: string,
-  output: string
-): boolean {
-  if (
-    level !== 'error' ||
-    !output.startsWith('jest-haste-map: @providesModule naming collision:')
-  ) {
-    return false;
-  }
-
-  const reactNativeNodeModulesPath = path.join(
-    projectRoot,
-    'node_modules',
-    'react-native',
-    'node_modules'
-  );
-  const reactNativeNodeModulesPattern = escapeRegExp(reactNativeNodeModulesPath);
-  const reactNativeNodeModulesCollisionRegex = new RegExp(
-    `Paths: ${reactNativeNodeModulesPattern}.+ collides with ${reactNativeNodeModulesPattern}.+`
-  );
-  return reactNativeNodeModulesCollisionRegex.test(output);
-}
-
-function _isIgnorableBugReportingExtraData(body: any[]) {
-  return body.length === 2 && body[0] === 'BugReporting extraData:';
-}
-
-function _isAppRegistryStartupMessage(body: any[]) {
-  return (
-    body.length === 1 &&
-    (/^Running application "main" with appParams:/.test(body[0]) ||
-      /^Running "main" with \{/.test(body[0]))
-  );
-}
-
-type ConsoleLogLevel = 'info' | 'warn' | 'error' | 'debug';
-
-function _handleDeviceLogs(projectRoot: string, deviceId: string, deviceName: string, logs: any) {
-  for (let i = 0; i < logs.length; i++) {
-    const log = logs[i];
-    let body = typeof log.body === 'string' ? [log.body] : log.body;
-    let { level } = log;
-
-    if (_isIgnorableBugReportingExtraData(body)) {
-      level = 'debug';
-    }
-    if (_isAppRegistryStartupMessage(body)) {
-      body = [`Running application on ${deviceName}.`];
-    }
-
-    const args = body.map((obj: any) => {
-      if (typeof obj === 'undefined') {
-        return 'undefined';
-      }
-      if (obj === 'null') {
-        return 'null';
-      }
-      if (typeof obj === 'string' || typeof obj === 'number' || typeof obj === 'boolean') {
-        return obj;
-      }
-      try {
-        return JSON.stringify(obj);
-      } catch (e) {
-        return obj.toString();
-      }
-    });
-    const logLevel =
-      level === 'info' || level === 'warn' || level === 'error' || level === 'debug'
-        ? (level as ConsoleLogLevel)
-        : 'info';
-    ProjectUtils.getLogger(projectRoot)[logLevel](
-      {
-        tag: 'device',
-        deviceId,
-        deviceName,
-        groupDepth: log.groupDepth,
-        shouldHide: log.shouldHide,
-        includesStack: log.includesStack,
-      },
-      ...args
-    );
-  }
-}
-export async function startReactNativeServerAsync({
-  projectRoot,
-  options = {},
-  exp = getConfig(projectRoot).exp,
-  verbose = true,
-}: {
-  projectRoot: string;
-  options: StartOptions;
-  exp?: ExpoConfig;
-  verbose?: boolean;
-}): Promise<void> {
-  assertValidProjectRoot(projectRoot);
-  await stopReactNativeServerAsync(projectRoot);
-  await Watchman.addToPathAsync(); // Attempt to fix watchman if it's hanging
-  await Watchman.unblockAndGetVersionAsync(projectRoot);
-
-  let packagerPort = await _getFreePortAsync(19001); // Create packager options
-
-  const customLogReporterPath: string = require.resolve(path.join(__dirname, '../build/reporter'));
-
-  // TODO: Bacon: Support .mjs (short-lived JS modules extension that some packages use)
-  const sourceExtsConfig = { isTS: true, isReact: true, isModern: false };
-  const sourceExts =
-    options.target === 'bare'
-      ? getBareExtensions([], sourceExtsConfig)
-      : getManagedExtensions([], sourceExtsConfig);
-
-  let packagerOpts: { [key: string]: any } = {
-    port: packagerPort,
-    customLogReporterPath,
-    sourceExts,
-  };
-
-  if (options.nonPersistent && Versions.lteSdkVersion(exp, '32.0.0')) {
-    packagerOpts.nonPersistent = true;
-  }
-
-  if (Versions.gteSdkVersion(exp, '33.0.0')) {
-    // starting with SDK 37, we bundle this plugin with the expo-asset package instead of expo,
-    // so check there first and fall back to expo if we can't find it in expo-asset
-    try {
-      packagerOpts.assetPlugins = resolveModule(
-        'expo-asset/tools/hashAssetFiles',
-        projectRoot,
-        exp
-      );
-    } catch (e) {
-      try {
-        packagerOpts.assetPlugins = resolveModule('expo/tools/hashAssetFiles', projectRoot, exp);
-      } catch (e) {
-        throw new Error(
-          'Unable to find the expo-asset package in the current project. Install it and try again.'
-        );
-      }
-    }
-  }
-
-  if (options.maxWorkers) {
-    packagerOpts['max-workers'] = options.maxWorkers;
-  }
-
-  if (!Versions.gteSdkVersion(exp, '16.0.0')) {
-    delete packagerOpts.customLogReporterPath;
-  }
-  const userPackagerOpts = exp.packagerOpts;
-
-  if (userPackagerOpts) {
-    // The RN CLI expects rn-cli.config.js's path to be absolute. We use the
-    // project root to resolve relative paths since that was the original
-    // behavior of the RN CLI.
-    if (userPackagerOpts.config) {
-      userPackagerOpts.config = path.resolve(projectRoot, userPackagerOpts.config);
-    }
-
-    // Provide a fallback if the value isn't given
-    const userSourceExts = userPackagerOpts.sourceExts ?? [];
-
-    packagerOpts = {
-      ...packagerOpts,
-      ...userPackagerOpts,
-      // In order to prevent people from forgetting to include the .expo extension or other things
-      // NOTE(brentvatne): we should probably do away with packagerOpts soon in favor of @expo/metro-config!
-      sourceExts: [...new Set([...packagerOpts.sourceExts, ...userSourceExts])],
-    };
-
-    if (userPackagerOpts.port !== undefined && userPackagerOpts.port !== null) {
-      packagerPort = userPackagerOpts.port;
-    }
-  }
-  const cliOpts = ['start'];
-  for (const [key, val] of Object.entries(packagerOpts)) {
-    // If the packager opt value is boolean, don't set
-    // --[opt] [value], just set '--opt'
-    if (val && typeof val === 'boolean') {
-      cliOpts.push(`--${key}`);
-    } else if (val) {
-      cliOpts.push(`--${key}`, val);
-    }
-  }
-
-  if (process.env.EXPO_DEBUG) {
-    cliOpts.push('--verbose');
-  }
-
-  if (options.reset) {
-    cliOpts.push('--reset-cache');
-  }
-
-  // Get the CLI path
-  const cliPath = resolveModule('react-native/local-cli/cli.js', projectRoot, exp);
-
-  // Run the copy of Node that's embedded in Electron by setting the
-  // ELECTRON_RUN_AS_NODE environment variable
-  // Note: the CLI script sets up graceful-fs and sets ulimit to 4096 in the
-  // child process
-  const packagerProcess = child_process.fork(cliPath, cliOpts, {
-    cwd: projectRoot,
-    env: {
-      ...process.env,
-      NODE_OPTIONS: process.env.METRO_NODE_OPTIONS,
-      REACT_NATIVE_APP_ROOT: projectRoot,
-      ELECTRON_RUN_AS_NODE: '1',
-    },
-    silent: true,
-  });
-  await ProjectSettings.setPackagerInfoAsync(projectRoot, {
-    packagerPort,
-    packagerPid: packagerProcess.pid,
-  }); // TODO: do we need this? don't know if it's ever called
-  process.on('exit', () => {
-    treekill(packagerProcess.pid);
-  });
-  if (!packagerProcess.stdout) {
-    throw new Error('Expected spawned process to have a stdout stream, but none was found.');
-  }
-  if (!packagerProcess.stderr) {
-    throw new Error('Expected spawned process to have a stderr stream, but none was found.');
-  }
-  packagerProcess.stdout.setEncoding('utf8');
-  packagerProcess.stderr.setEncoding('utf8');
-  packagerProcess.stdout.pipe(split()).on('data', data => {
-    if (verbose) {
-      _logPackagerOutput(projectRoot, 'info', data);
-    }
-  });
-  packagerProcess.stderr.on('data', data => {
-    if (verbose) {
-      _logPackagerOutput(projectRoot, 'error', data);
-    }
-  });
-  const exitPromise = new Promise<void>((resolve, reject) => {
-    packagerProcess.once('exit', async code => {
-      ProjectUtils.logDebug(projectRoot, 'expo', `Metro Bundler process exited with code ${code}`);
-      if (code) {
-        reject(new Error(`Metro Bundler process exited with code ${code}`));
-      } else {
-        resolve();
-      }
-      try {
-        await ProjectSettings.setPackagerInfoAsync(projectRoot, {
-          packagerPort: null,
-          packagerPid: null,
-        });
-      } catch (e) {}
-    });
-  });
-  const packagerUrl = await UrlUtils.constructBundleUrlAsync(projectRoot, {
-    urlType: 'http',
-    hostType: 'localhost',
-  });
-  await Promise.race([_waitForRunningAsync(projectRoot, `${packagerUrl}/status`), exitPromise]);
-}
-
-export async function stopReactNativeServerAsync(projectRoot: string): Promise<void> {
-  assertValidProjectRoot(projectRoot);
-  const packagerInfo = await ProjectSettings.readPackagerInfoAsync(projectRoot);
-  if (!packagerInfo.packagerPort || !packagerInfo.packagerPid) {
-    ProjectUtils.logDebug(projectRoot, 'expo', `No packager found for project at ${projectRoot}.`);
-    return;
-  }
-  ProjectUtils.logDebug(
-    projectRoot,
-    'expo',
-    `Killing packager process tree: ${packagerInfo.packagerPid}`
-  );
-  try {
-    await treekillAsync(packagerInfo.packagerPid, 'SIGKILL');
-  } catch (e) {
-    ProjectUtils.logDebug(projectRoot, 'expo', `Error stopping packager process: ${e.toString()}`);
-  }
-  await ProjectSettings.setPackagerInfoAsync(projectRoot, {
-    packagerPort: null,
-    packagerPid: null,
-  });
-}
-
-export async function startExpoServerAsync(
-  projectRoot: string,
-  options: StartOptions
-): Promise<void> {
-  assertValidProjectRoot(projectRoot);
-  await stopExpoServerAsync(projectRoot);
-  const app = express();
-  app.use(
-    express.json({
-      limit: '10mb',
-    })
-  );
-  app.use(
-    express.urlencoded({
-      limit: '10mb',
-      extended: true,
-    })
-  );
-  if (
-    (ConnectionStatus.isOffline()
-      ? await Doctor.validateWithoutNetworkAsync(projectRoot)
-      : await Doctor.validateWithNetworkAsync(projectRoot)) === Doctor.FATAL
-  ) {
-    throw new Error(`Couldn't start project. Please fix the errors and restart the project.`);
-  }
-  // Serve the manifest.
-  const manifestHandler = getManifestHandler(projectRoot);
-  app.get('/', manifestHandler);
-  app.get('/manifest', manifestHandler);
-  app.get('/index.exp', manifestHandler);
-  app.post('/logs', async (req, res) => {
-    try {
-      const deviceId = req.get('Device-Id');
-      const deviceName = req.get('Device-Name');
-      if (deviceId && deviceName && req.body) {
-        _handleDeviceLogs(projectRoot, deviceId, deviceName, req.body);
-      }
-    } catch (e) {
-      ProjectUtils.logError(projectRoot, 'expo', `Error getting device logs: ${e} ${e.stack}`);
-    }
-    res.send('Success');
-  });
-  app.post('/shutdown', async (req, res) => {
-    server.close();
-    res.send('Success');
-  });
-  const expRc = await readExpRcAsync(projectRoot);
-  const expoServerPort = expRc.manifestPort ? expRc.manifestPort : await _getFreePortAsync(19000);
-  await ProjectSettings.setPackagerInfoAsync(projectRoot, {
-    expoServerPort,
-  });
-  let server = app.listen(expoServerPort, () => {
-    const info = server.address() as AddressInfo;
-    const host = info.address;
-    const port = info.port;
-    ProjectUtils.logDebug(projectRoot, 'expo', `Local server listening at http://${host}:${port}`);
-  });
-}
-
-async function stopExpoServerAsync(projectRoot: string): Promise<void> {
-  assertValidProjectRoot(projectRoot);
-  const packagerInfo = await ProjectSettings.readPackagerInfoAsync(projectRoot);
-  if (packagerInfo && packagerInfo.expoServerPort) {
-    try {
-      await axios.request({
-        method: 'post',
-        url: `http://127.0.0.1:${packagerInfo.expoServerPort}/shutdown`,
-      });
-    } catch (e) {}
-  }
-  await ProjectSettings.setPackagerInfoAsync(projectRoot, {
-    expoServerPort: null,
-  });
-}
-
-async function startDevServerAsync(projectRoot: string, startOptions: StartOptions) {
-  assertValidProjectRoot(projectRoot);
-
-  const port = startOptions.devClient
-    ? Number(process.env.RCT_METRO_PORT) || 8081
-    : await _getFreePortAsync(19000);
-  await ProjectSettings.setPackagerInfoAsync(projectRoot, {
-    expoServerPort: port,
-    packagerPort: port,
-  });
-
-  const options: MetroDevServerOptions = {
-    port,
-    logger: ProjectUtils.getLogger(projectRoot),
-  };
-  if (startOptions.reset) {
-    options.resetCache = true;
-  }
-  if (startOptions.maxWorkers != null) {
-    options.maxWorkers = startOptions.maxWorkers;
-  }
-  if (startOptions.target) {
-    // EXPO_TARGET is used by @expo/metro-config to determine the target when getDefaultConfig is
-    // called from metro.config.js.
-    process.env.EXPO_TARGET = startOptions.target;
-  }
-
-  const { middleware } = await runMetroDevServerAsync(projectRoot, options);
-  middleware.use(getManifestHandler(projectRoot));
-}
-
 export async function setOptionsAsync(
   projectRoot: string,
   options: {
@@ -1829,7 +1306,7 @@ export async function startAsync(
     await startDevServerAsync(projectRoot, options);
     DevSession.startSession(projectRoot, exp, 'native');
   } else {
-    await startExpoServerAsync(projectRoot, options);
+    await startExpoServerAsync(projectRoot);
     await startReactNativeServerAsync({ projectRoot, exp, options, verbose });
     DevSession.startSession(projectRoot, exp, 'native');
   }
@@ -1899,4 +1376,11 @@ export async function stopAsync(projectDir: string): Promise<void> {
   }
 }
 
-export { startTunnelsAsync, stopTunnelsAsync };
+export {
+  startTunnelsAsync,
+  stopTunnelsAsync,
+  startExpoServerAsync,
+  StartOptions,
+  startReactNativeServerAsync,
+  stopReactNativeServerAsync,
+};
